Fixed gaiaid type as bigInt in migrations

+++ data/migrations/20190625185034_mock_tables.js
@@ -22,7 +22,7 @@ exports.up = function(knex, Promise) {
         tbl
             .string('twomass', 255);
         tbl
-            .integer('gaiaid');
+            .bigInteger('gaiaid');
         tbl
             .string('sector');
         tbl
